[Decomp][Spell] Cohesion split: extract EffectDummy handler to SpellEffectDummy.cpp
